requirements.txt
alembic==1.13.1
Flask==3.0.3
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
tensorflow==2.15.0
python-dotenv==1.0.1
numpy==1.26.2
Pillow==10.1.0